# Update murmurbytes and benchmarks