--- a/murmur_test.go
+++ b/murmur_test.go
@@ -83,6 +83,7 @@ func TestMurmurStringZero(t *testing.T) {
 }
 
 func randString(n int) string {
+	rand.Seed(10)
 	letterRunes := []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
 	b := make([]rune, n)
 	for i := range b {
@@ -92,6 +93,15 @@ func randString(n int) string {
 }
 
 // Benchmarks
+func benchmarkMurmurBytes(b *testing.B, input [][]byte) {
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		for _, x := range input {
+			MurmurBytes(x)
+		}
+	}
+}
+
 func benchmarkMurmur64(b *testing.B, input []uint64) {
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
@@ -120,7 +130,19 @@ func benchmarkHash32(b *testing.B, input []string) {
 	}
 }
 
+func Benchmark100MurmurBytes(b *testing.B) {
+	rand.Seed(10)
+	input := make([][]byte, 100)
+	for i := 0; i < 100; i++ {
+		x := make([]byte, 1000)
+		rand.Read(x)
+		input[i] = x
+	}
+	benchmarkMurmurBytes(b, input)
+}
+
 func Benchmark100Murmur64(b *testing.B) {
+	rand.Seed(10)
 	input := make([]uint64, 100)
 	for i := 0; i < 100; i++ {
 		input[i] = uint64(rand.Int63())
@@ -129,6 +151,7 @@ func Benchmark100Murmur64(b *testing.B) {
 }
 
 func Benchmark100MurmurString(b *testing.B) {
+	rand.Seed(10)
 	input := make([]string, 100)
 	for i := 0; i < 100; i++ {
 		input[i] = randString((i % 15) + 5)
@@ -137,6 +160,7 @@ func Benchmark100MurmurString(b *testing.B) {
 }
 
 func Benchmark100Hash32(b *testing.B) {
+	rand.Seed(10)
 	input := make([]string, 100)
 	for i := 0; i < 100; i++ {
 		input[i] = randString((i % 15) + 5)