This PR removes calls to `unsafe.Pointer` from `MurmuBytes` in order to be compatible with [go 1.14](https://golang.org/doc/go1.14#compiler). It also adds benchmark tests for this function. Results of running the new tests added in this PR against master using `go test -bench=Murmur`:

```
$ benchcmp old new
benchmark                      old ns/op     new ns/op     delta
Benchmark100MurmurBytes-8      28618         33227         +16.11%
Benchmark100Murmur64-8         347           367           +5.76%
Benchmark100MurmurString-8     833           928           +11.40%
BenchmarkMurmurStringBig-8     30933217      37862688      +22.40%
```

As we can see, this change adds significant overhead to the functions that call `MurmurBytes`. (The Murmur64 benchmark can likely be safely ignored, since this PR does not touch its functionality).